almost finished extra pcb
The pcb is designed to be single layerd for simplisity, but there are still two unconnected lines, GND and 5V are now both in two seperate sections. Can connect via an wire or update the PCB. Will not spend a lot of time on it now for it is not shure we will use an PCB or just a board with lots of wires on it.

Signed-off-by: Abel Tim <git@abeltim.com>